requirements/docs.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra docs -o requirements/docs.txt -c requirements/requirements.txt pyproject.toml
alabaster==0.7.16
    # via sphinx
annotated-types==0.7.0
    # via
    #   -c requirements/requirements.txt
    #   pydantic
anyio==4.4.0
    # via
    #   -c requirements/requirements.txt
    #   httpx
    #   starlette
    #   watchfiles
attrs==24.2.0
    # via
    #   cattrs
    #   lsprotocol
autodoc-pydantic==2.2.0
    # via hirundo (pyproject.toml)
babel==2.15.0
    # via sphinx
backports-tarfile==1.2.0
    # via
    #   -c requirements/requirements.txt
    #   jaraco-context
beautifulsoup4==4.12.3
    # via furo
cattrs==24.1.2
    # via
    #   lsprotocol
    #   pygls
certifi==2024.7.4
    # via
    #   -c requirements/requirements.txt
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via
    #   -c requirements/requirements.txt
    #   requests
click==8.1.7
    # via
    #   -c requirements/requirements.txt
    #   sphinx-click
    #   typer
    #   uvicorn
colorama==0.4.6
    # via sphinx-autobuild
docutils==0.21.2
    # via
    #   -c requirements/requirements.txt
    #   readme-renderer
    #   sphinx
    #   sphinx-click
esbonio==0.16.5
    # via hirundo (pyproject.toml)
exceptiongroup==1.2.2
    # via
    #   -c requirements/requirements.txt
    #   anyio
    #   cattrs
furo==2024.7.18
    # via hirundo (pyproject.toml)
h11==0.14.0
    # via
    #   -c requirements/requirements.txt
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via
    #   -c requirements/requirements.txt
    #   httpx
httpx==0.27.0
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
httpx-sse==0.4.0
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
idna==3.7
    # via
    #   -c requirements/requirements.txt
    #   anyio
    #   httpx
    #   requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.0.0
    # via
    #   -c requirements/requirements.txt
    #   keyring
    #   sphinx
    #   twine
jaraco-classes==3.4.0
    # via
    #   -c requirements/requirements.txt
    #   keyring
jaraco-context==5.3.0
    # via
    #   -c requirements/requirements.txt
    #   keyring
jaraco-functools==4.0.1
    # via
    #   -c requirements/requirements.txt
    #   keyring
jinja2==3.1.4
    # via sphinx
keyring==25.2.1
    # via
    #   -c requirements/requirements.txt
    #   twine
lsprotocol==2023.0.1
    # via pygls
markdown-it-py==3.0.0
    # via
    #   -c requirements/requirements.txt
    #   rich
markupsafe==3.0.2
    # via
    #   hirundo (pyproject.toml)
    #   jinja2
mdurl==0.1.2
    # via
    #   -c requirements/requirements.txt
    #   markdown-it-py
more-itertools==10.3.0
    # via
    #   -c requirements/requirements.txt
    #   jaraco-classes
    #   jaraco-functools
nh3==0.2.18
    # via
    #   -c requirements/requirements.txt
    #   readme-renderer
packaging==24.1
    # via sphinx
pkginfo==1.10.0
    # via
    #   -c requirements/requirements.txt
    #   twine
platformdirs==4.3.6
    # via esbonio
pydantic==2.8.2
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
    #   autodoc-pydantic
    #   pydantic-settings
pydantic-core==2.20.1
    # via
    #   -c requirements/requirements.txt
    #   pydantic
pydantic-settings==2.3.4
    # via autodoc-pydantic
pygls==1.3.1
    # via esbonio
pygments==2.18.0
    # via
    #   -c requirements/requirements.txt
    #   furo
    #   readme-renderer
    #   rich
    #   sphinx
pyspellchecker==0.8.1
    # via esbonio
python-dotenv==1.0.1
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
    #   pydantic-settings
pyyaml==6.0.1
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
readme-renderer==44.0
    # via
    #   -c requirements/requirements.txt
    #   twine
requests==2.32.3
    # via
    #   -c requirements/requirements.txt
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==1.0.0
    # via
    #   -c requirements/requirements.txt
    #   twine
rfc3986==2.0.0
    # via
    #   -c requirements/requirements.txt
    #   twine
rich==13.7.1
    # via
    #   -c requirements/requirements.txt
    #   twine
    #   typer
shellingham==1.5.4
    # via
    #   -c requirements/requirements.txt
    #   typer
sniffio==1.3.1
    # via
    #   -c requirements/requirements.txt
    #   anyio
    #   httpx
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==7.4.7
    # via
    #   hirundo (pyproject.toml)
    #   autodoc-pydantic
    #   esbonio
    #   furo
    #   sphinx-autobuild
    #   sphinx-basic-ng
    #   sphinx-click
    #   sphinx-multiversion
sphinx-autobuild==2024.4.16
    # via hirundo (pyproject.toml)
sphinx-basic-ng==1.0.0b2
    # via furo
sphinx-click==6.0.0
    # via hirundo (pyproject.toml)
sphinx-multiversion==0.2.4
    # via hirundo (pyproject.toml)
sphinxcontrib-applehelp==1.0.8
    # via sphinx
sphinxcontrib-devhelp==1.0.6
    # via sphinx
sphinxcontrib-htmlhelp==2.0.6
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.8
    # via sphinx
sphinxcontrib-serializinghtml==1.1.10
    # via sphinx
stamina==24.2.0
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
starlette==0.41.2
    # via
    #   hirundo (pyproject.toml)
    #   sphinx-autobuild
tenacity==8.5.0
    # via
    #   -c requirements/requirements.txt
    #   stamina
tomli==2.0.1
    # via sphinx
tqdm==4.66.5
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
twine==5.1.1
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
typer==0.12.3
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
types-pyyaml==6.0.12.20240311
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
types-requests==2.32.0.20240712
    # via
    #   -c requirements/requirements.txt
    #   hirundo (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   -c requirements/requirements.txt
    #   anyio
    #   cattrs
    #   pydantic
    #   pydantic-core
    #   stamina
    #   starlette
    #   typer
    #   uvicorn
urllib3==2.2.2
    # via
    #   -c requirements/requirements.txt
    #   requests
    #   twine
    #   types-requests
uvicorn==0.30.3
    # via sphinx-autobuild
watchfiles==0.22.0
    # via sphinx-autobuild
websockets==12.0
    # via sphinx-autobuild
zipp==3.19.2
    # via
    #   -c requirements/requirements.txt
    #   importlib-metadata
